Modularize validate phone number
This enables us to just import the merge_validate_phone instead of loading the entire Mailchimp plugin
